Try to reconnect infinitely after a timeout
If a connection times out after it connected successfully, tsclientlib
waits for 10s and then tries to reconnect. As long as no response is
received (i.e. reconnecting fails with a timeout), it tries again.
When another error is received, e.g. socket cannot be opened or the
server responds with an error, connecting is aborted.